refactor(mongo-journal-model): rename 'journal-model' to 'mongo-journal-model'
